Reestructure edit_commands
