+++ src/pycture/commands/edit_commands/__init__.py
@@ -1,0 +1,3 @@
+from .transform_by_linear_segments import TransformByLinearSegments
+from .edit_brightness import EditBrightness
+from .convert_to_gray_scale import ConvertToGrayScale
