+++ src/pycture/commands/edit_commands/edit_brightness.py
@@ -1,0 +1,10 @@
+from PyQt5.QtWidgets import QWidget, QMainWindow
+from ..command import Command
+
+
+class EditBrightness(Command):
+    def __init__(self, parent: QWidget):
+        super().__init__(parent, "Brightness")
+
+    def execute(self, main_window: QMainWindow):
+        print("Edits brightness")
