+++ src/pycture/dialogs/__init__.py
@@ -1,2 +1,3 @@
 from .yes_cancel_notification import YesCancelNotification
 from .notification import Notification
+from .input_dialogs import SegmentsInput
